Pre-Gold
About to submit gold.

Game now builds without any warnings.

+++ enemytype2.h
@@ -39,7 +39,7 @@ public:
 
 	int GetRadius() override
 	{
-		return (m_animatedSprite->GetFrameWidth() / 2) * m_animatedSprite->GetScale() * 0.3f;
+		return static_cast<int>((m_animatedSprite->GetFrameWidth() / 2) * m_animatedSprite->GetScale() * 0.3f);
 	}
 };
